+++ CHANGELOG.md
@@ -6,9 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - 2020-XX-XX
 ### Added
-- a
+- [Plotly for Rust Book](https://igiagkiozis.github.io/plotly/)
 ### Changed
-- c
+- `Layout` complete.
 ### Fixed
 - f